+++ src/Entities/PayoutTransaction.php
@@ -1,0 +1,27 @@
+<?php
+
+namespace Payrex\Entities;
+
+class PayoutTransaction extends \Payrex\Entities\BaseEntity
+{
+    public $id;
+    public $amount;
+    public $payout_id;
+    public $transaction_type;
+    public $transaction_id;
+    public $created_at;
+    public $updated_at;
+
+    public function __construct($apiResource)
+    {
+        $data = $apiResource->data;
+
+        $this->id = $data['id'];
+        $this->amount = $data['amount'];
+        $this->payout_id = $data['payout_id'];
+        $this->transaction_type = $data['transaction_type'];
+        $this->transaction_id = $data['transaction_id'];
+        $this->created_at = $data['created_at'];
+        $this->updated_at = $data['updated_at'];
+    }
+}