+++ src/Services/PayoutService.php
@@ -1,0 +1,23 @@
+<?php
+
+namespace Payrex\Services;
+
+class PayoutService extends \Payrex\Services\BaseService {
+    const URI = '/payouts';
+
+    public function list_transactions($id, $params = []) {
+        $apiResponse = $this->httpClient->request([
+            'method' => 'GET',
+            'url'    => "{$this->client->apiBaseUrl}" . self::URI . "/{$id}/transactions",
+            'params' => $params
+        ]);
+
+        foreach ($apiResponse->data['data'] as $key => $value) {
+            $apiResponse->data['data'][$key] = new \Payrex\Entities\PayoutTransaction(
+                new \Payrex\ApiResource($value)
+            );
+        }
+
+        return new \Payrex\Entities\Listing($apiResponse);
+    }
+}