+++ src/Services/ServiceFactory.php
@@ -11,6 +11,7 @@ class ServiceFactory
         'customerSessions'          => \Payrex\Services\CustomerSessionService::class,
         'customers'                 => \Payrex\Services\CustomerService::class,
         'payments'                  => \Payrex\Services\PaymentService::class,
+        'payouts'                   => \Payrex\Services\PayoutService::class,
         'paymentIntents'            => \Payrex\Services\PaymentIntentService::class,
         'refunds'                   => \Payrex\Services\RefundService::class,
         'webhooks'                  => \Payrex\Services\WebhookService::class